requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers==3.14.0  # Pour gérer les CORS
gunicorn==20.1.0  # Serveur WSGI pour la production
Pillow==10.0.0  # Pour le traitement d'images
psycopg2-binary==2.9.7  # Pour la connexion à PostgreSQL
dj-database-url==0.5.0
